+++ src/main/java/com/coupop/fcfscoupon/fcfsissue/model/FcfsIssuePolicy.java
@@ -1,4 +1,4 @@
-package com.coupop.fcfscoupon.model;
+package com.coupop.fcfscoupon.fcfsissue.model;
 
 import java.time.LocalTime;
