+++ src/main/java/com/coupop/fcfscoupon/fcfsissue/model/FcfsIssueRepository.java
@@ -1,4 +1,4 @@
-package com.coupop.fcfscoupon.model;
+package com.coupop.fcfscoupon.fcfsissue.model;
 
 public interface FcfsIssueRepository {
